New function to return books with certain genre
Added a new function that searches the collection for a certain genre
and all books that contain that genre. Have to put the collection of
books into a list to return any meaningful data.

This then stop the code for dividing the books into pages of 12 each...
Have to create a workaround for my pagination that can count number
of books in the list.