Changed to have dropdown only present for users
Changed the way the navbar will be displayed for users and non-users
allowing the login/register sector to be immediately visible
and for logged in users to only see the dropdown.

This is to prevent new users having to first click on the dropdown
to understand that they need to log in or register to get the full
capabilities of the site.