Added a char count for isbn & synopsis
Adding the character counter for both holds certain weight for me and
the user. For the ISBN I want it it only been the 10 digit ISBN due
to it having the effect of linking to amazons page. But I cannot
restrict input to 10 characters on type=number.

Synopsis has maxlength of 5000 chars and counter to let users know if
they are close or not.